projects/microservice/exporter/pyproject.toml
[tool.poetry]
name = "exporter"
version = "0.0.1"
description = "DeepClean production IaaS export application"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
tensorrt = "8.5.1.7"
flask = "^2.2"

torch = {version = "^1.10", source = "torch"}
ml4gw = {path = "../../../ml4gw", develop = true}

typeo = {path = "../../../typeo", develop = true}
"deepclean.architectures" = {path = "../../../libs/architectures", develop = true}
"deepclean.export" = {path = "../../../libs/export", develop = true}
"deepclean.logging" = {path = "../../../libs/logging", develop = true}
"deepclean.utils" = {path = "../../../libs/utils", develop = true}

[tool.poetry.dependencies."hermes.quiver"]
path = "../../../hermes/hermes/hermes.quiver"
develop = true
extras = ["torch", "tensortt"]

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
